Deduplicate smarter.
The logic here is pretty simple. I track the number of matching tasks for the task in question
vs. how many times we have skipped the task already. I keep skipping until these numbers are
equal. Any entries after that are considered new. Tested this, and it works. It's general
enough that I added it to \TogglSync\VastSyncer.